requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
requests==2.32.4
python-dotenv==1.1.1
jsonschema==4.25.0
jinja2==3.1.6
